Dear 132 fix workkind per day count chart (#24)
* DEAR-132 fix  workkind count per day, 1 user

* DEAR-132 fix  workkind count per day, 1 user, different timeranges

* DEAR-132 fix  workkind count per day, 1 user, different timeranges, per sprint

* DEAR-132 Add testdata workkind count

* DEAR-132 clean up, remove excluding current user

* DEAR-132 add sprint to insights, wip

* DEAR-132 fix 1

* DEAR-132 fix 2

* DEAR-132 convert dto responses to integers

* DEAR-132 clean up

* DEAR-132 disable test

* DEAR-132 disable test

* DEAR-132 add github username

* DEAR-132 update github user name

* DEAR-132 add teams with sprints

* DEAR-132 add todo

+++ src/main/java/ch/fhnw/deardevbackend/dto/insights/HappinessInsightDTO.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class HappinessInsightDTO {
     private String day;
-    private Double userAverage;
-    private Double teamAverage;
+    private Integer userAverage;
+    private Integer teamAverage;
 }
